requirements.txt
groq
python-dotenv
PyPDF2
flask